+++ tests/test_daemon.py
@@ -327,340 +327,6 @@ class TestOspdOpenvas(TestCase):
 
         self.assertTrue(w.sudo_available)
 
-    def test_get_custom_xml(self):
-        out = (
-            '<custom>'
-            '<required_ports>Services/www, 80</required_ports>'
-            '<category>3</category>'
-            '<excluded_keys>Settings/disable_cgi_scanning</excluded_keys>'
-            '<family>Product detection</family>'
-            '<filename>mantis_detect.nasl</filename>'
-            '<timeout>0</timeout>'
-            '</custom>'
-        )
-        w = DummyDaemon()
-
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        res = w.get_custom_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', vt.get('custom')
-        )
-        self.assertEqual(len(res), len(out))
-
-    def test_get_custom_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        custom = {'a': "\u0006"}
-        w.get_custom_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', custom=custom
-        )
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_severities_xml(self):
-        w = DummyDaemon()
-
-        out = (
-            '<severities>'
-            '<severity type="cvss_base_v2">'
-            '<value>AV:N/AC:L/Au:N/C:N/I:N/A:N</value>'
-            '<origin>Greenbone</origin>'
-            '<date>1237458156</date>'
-            '</severity>'
-            '</severities>'
-        )
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        severities = vt.get('severities')
-        res = w.get_severities_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', severities
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_severities_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        sever = {'severity_base_vector': "\u0006"}
-        w.get_severities_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', severities=sever
-        )
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_params_xml(self):
-        w = DummyDaemon()
-        out = (
-            '<params>'
-            '<param type="checkbox" id="2">'
-            '<name>Do not randomize the  order  in  which ports are '
-            'scanned</name>'
-            '<default>no</default>'
-            '</param>'
-            '<param type="entry" id="1">'
-            '<name>Data length :</name>'
-            '</param>'
-            '</params>'
-        )
-
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        params = vt.get('vt_params')
-        res = w.get_params_vt_as_xml_str('1.3.6.1.4.1.25623.1.0.100061', params)
-
-        self.assertEqual(len(res), len(out))
-
-    def test_get_params_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        params = {
-            '1': {
-                'id': '1',
-                'type': 'entry',
-                'default': '\u0006',
-                'name': 'dns-fuzz.timelimit',
-                'description': 'Description',
-            }
-        }
-        w.get_params_vt_as_xml_str('1.3.6.1.4.1.25623.1.0.100061', params)
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_refs_xml(self):
-        w = DummyDaemon()
-
-        out = '<refs><ref type="url" id="http://www.mantisbt.org/"/></refs>'
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        refs = vt.get('vt_refs')
-        res = w.get_refs_vt_as_xml_str('1.3.6.1.4.1.25623.1.0.100061', refs)
-
-        self.assertEqual(res, out)
-
-    def test_get_dependencies_xml(self):
-        w = DummyDaemon()
-
-        out = (
-            '<dependencies>'
-            '<dependency vt_id="1.3.6.1.4.1.25623.1.2.3.4"/>'
-            '<dependency vt_id="1.3.6.1.4.1.25623.4.3.2.1"/>'
-            '</dependencies>'
-        )
-        dep = ['1.3.6.1.4.1.25623.1.2.3.4', '1.3.6.1.4.1.25623.4.3.2.1']
-        res = w.get_dependencies_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', dep
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_dependencies_xml_missing_dep(self):
-        w = DummyDaemon()
-
-        out = (
-            '<dependencies>'
-            '<dependency vt_id="1.3.6.1.4.1.25623.1.2.3.4"/>'
-            '</dependencies>'
-        )
-        dep = ['1.3.6.1.4.1.25623.1.2.3.4', 'file_name.nasl']
-        res = w.get_dependencies_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', dep
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_dependencies_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.error = Mock()
-
-        dep = ["\u0006"]
-        w.get_dependencies_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', vt_dependencies=dep
-        )
-
-        assert_called_once(logging.Logger.error)
-
-    def test_get_ctime_xml(self):
-        w = DummyDaemon()
-
-        out = '<creation_time>1237458156</creation_time>'
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        ctime = vt.get('creation_time')
-        res = w.get_creation_time_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', ctime
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_ctime_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        ctime = '\u0006'
-        w.get_creation_time_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', vt_creation_time=ctime
-        )
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_mtime_xml(self):
-        w = DummyDaemon()
-
-        out = '<modification_time>1533906565</modification_time>'
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        mtime = vt.get('modification_time')
-        res = w.get_modification_time_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', mtime
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_mtime_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        mtime = '\u0006'
-        w.get_modification_time_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', mtime
-        )
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_summary_xml(self):
-        w = DummyDaemon()
-
-        out = '<summary>some summary</summary>'
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        summary = vt.get('summary')
-        res = w.get_summary_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', summary
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_summary_xml_failed(self):
-        w = DummyDaemon()
-
-        summary = '\u0006'
-        logging.Logger.warning = Mock()
-        w.get_summary_vt_as_xml_str('1.3.6.1.4.1.25623.1.0.100061', summary)
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_impact_xml(self):
-        w = DummyDaemon()
-
-        out = '<impact>some impact</impact>'
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        impact = vt.get('impact')
-        res = w.get_impact_vt_as_xml_str('1.3.6.1.4.1.25623.1.0.100061', impact)
-
-        self.assertEqual(res, out)
-
-    def test_get_impact_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        impact = '\u0006'
-        w.get_impact_vt_as_xml_str('1.3.6.1.4.1.25623.1.0.100061', impact)
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_insight_xml(self):
-        w = DummyDaemon()
-
-        out = '<insight>some insight</insight>'
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        insight = vt.get('insight')
-        res = w.get_insight_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', insight
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_insight_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        insight = '\u0006'
-        w.get_insight_vt_as_xml_str('1.3.6.1.4.1.25623.1.0.100061', insight)
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_solution_xml(self):
-        w = DummyDaemon()
-
-        out = (
-            '<solution type="WillNotFix" method="DebianAPTUpgrade">'
-            'some solution'
-            '</solution>'
-        )
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        solution = vt.get('solution')
-        solution_type = vt.get('solution_type')
-        solution_method = vt.get('solution_method')
-
-        res = w.get_solution_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061',
-            solution,
-            solution_type,
-            solution_method,
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_solution_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        solution = '\u0006'
-        w.get_solution_vt_as_xml_str('1.3.6.1.4.1.25623.1.0.100061', solution)
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_detection_xml(self):
-        w = DummyDaemon()
-
-        out = '<detection qod_type="remote_banner"/>'
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        detection_type = vt.get('qod_type')
-
-        res = w.get_detection_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', qod_type=detection_type
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_detection_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        detection = '\u0006'
-        w.get_detection_vt_as_xml_str('1.3.6.1.4.1.25623.1.0.100061', detection)
-
-        assert_called_once(logging.Logger.warning)
-
-    def test_get_affected_xml(self):
-        w = DummyDaemon()
-        out = '<affected>some affection</affected>'
-        vt = w.VTS['1.3.6.1.4.1.25623.1.0.100061']
-        affected = vt.get('affected')
-
-        res = w.get_affected_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', affected=affected
-        )
-
-        self.assertEqual(res, out)
-
-    def test_get_affected_xml_failed(self):
-        w = DummyDaemon()
-        logging.Logger.warning = Mock()
-
-        affected = "\u0006" + "affected"
-        w.get_affected_vt_as_xml_str(
-            '1.3.6.1.4.1.25623.1.0.100061', affected=affected
-        )
-
-        assert_called_once(logging.Logger.warning)
-
     @patch('ospd_openvas.daemon.Path.exists')
     @patch('ospd_openvas.daemon.Path.open')
     def test_get_feed_info(
